No more $rootScope nor $apply and fix bug
There's no need for $apply nor for $rootScope as a consequence, so they
got removed.

There was a bug while using angular.fromJson() was creating an
"unexpected end of input". There's no need, the XHR object has already
several response types and now we are pushing them as a response.